Move Awards types into Seasons namespace
Refactor domain: relocate award-related types from Neba.Domain.Awards to Neba.Domain.Seasons. Updated namespaces across domain classes, AppDbContext, EF Core configuration classes, migration designer files and model snapshot, plus corresponding test and test-factory files to match the new namespace. Ensures award entities are organized under Seasons and keeps EF migrations and references consistent.
