+++ src/Neba.Infrastructure/Database/Configurations/HighAverageAwardConfiguration.cs
@@ -1,7 +1,7 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-using Neba.Domain.Awards;
+using Neba.Domain.Seasons;
 
 namespace Neba.Infrastructure.Database.Configurations;
