+++ tests/Neba.Domain.Tests/Seasons/HighBlockAwardTests.cs
@@ -1,4 +1,4 @@
-using Neba.Domain.Awards;
+using Neba.Domain.Seasons;
 using Neba.Domain.Bowlers;
 using Neba.TestFactory.Attributes;
